+++ CHANGELOG.md
@@ -29,6 +29,7 @@ That's all, matcher will stop when detects next line started with `###` match
 * Fixed some bugs on particle menu
 * Fixed when plugins scoreboards not re-appeared after game end
 * Removed temp added scoreboard rewards
+* Added new message In-Game.Floor-Item-Blacklisted
 
 ### 4.5.0 Release (21.03.2021 - 13.04.2021)
 * Automatically disable bossbar support on 1.8 to prevent issue if bossbar is enabled on config.yml